Move Phase 2 files to Phase 2 directory
