+++ packages/core/conductor-tasks/bichard_phase_2/persistPhase2.integration.test.ts
@@ -13,7 +13,7 @@ import insertErrorListRecord from "../../lib/database/insertErrorListRecord"
 import insertErrorListTriggers from "../../lib/database/insertErrorListTriggers"
 import errorPaths from "../../lib/exceptions/errorPaths"
 import generateMockPhase1Result from "../../phase1/tests/helpers/generateMockPhase1Result"
-import generateMockPhase2Result from "../../phase1/tests/helpers/generateMockPhase2Result"
+import generateMockPhase2Result from "../../phase2/tests/helpers/generateMockPhase2Result"
 import { Phase2ResultType } from "../../phase2/types/Phase2Result"
 import ResolutionStatus from "../../types/ResolutionStatus"
 import persistPhase2 from "./persistPhase2"
