+++ packages/core/phase2/tests/fixtures/helpers/generateFakePncUpdateDataset.ts
@@ -1,8 +1,8 @@
 import { dateReviver } from "@moj-bichard7/common/axiosDateTransformer"
 import fs from "fs"
 import merge from "lodash.merge"
-import type { PartialPncUpdateDataset } from "../../../../phase1/tests/helpers/PartialPncUpdateDataset"
 import type { PncUpdateDataset } from "../../../../types/PncUpdateDataset"
+import type { PartialPncUpdateDataset } from "../../helpers/PartialPncUpdateDataset"
 
 const generateFakePncUpdateDataset = (overrides: PartialPncUpdateDataset = { PncOperations: [] }): PncUpdateDataset => {
   const exampleAho = fs.readFileSync("./phase1/tests/fixtures/exampleAho.json").toString()
