+++ packages/core/phase2/tests/helpers/PartialPncUpdateDataset.ts
@@ -1,5 +1,5 @@
 import type { z } from "zod"
-import pncUpdateDatasetSchema from "../../../phase2/schemas/pncUpdateDataset"
+import pncUpdateDatasetSchema from "../../schemas/pncUpdateDataset"
 
 const partialPncUpdateDataset = pncUpdateDatasetSchema.deepPartial()
 export type PartialPncUpdateDataset = z.infer<typeof partialPncUpdateDataset>
